+++ src/cpp/MaterialDot.hpp
@@ -14,11 +14,11 @@
 class MaterialDot {
 private:
     /// Coordinates
-    std::array<float, 3> coordinates;
+    std::array<double, 3> coordinates;
     /// Velocity
-    std::array<float, 3> velocity;
+    std::array<double, 3> velocity;
     /// Mass
-    float m = 1;
+    double m = 1;
 public:
     /**
      * MaterialDot object
@@ -27,46 +27,46 @@ public:
      * @param z means Z-coordinate
      * @param m means mass of a dot
      */
-    explicit MaterialDot(float x=0, float y=0, float z=0, float m=1);
+    explicit MaterialDot(double x=0, double y=0, double z=0, double m=1);
 
     /**
     * MaterialDot object
-    * @param coords array of floats with size 3
+    * @param coords array of doubles with size 3
     * that will be considered as initial coordinates
     * @param m means mass of a dot
     */
-    explicit MaterialDot(float* coords, float m=1);
+    explicit MaterialDot(double* coords, double m=1);
 
     /**
      * Get material dot coordinates as an array
-     * @return Array of floats with copy of x, y, z coordinates respectively
+     * @return Array of doubles with copy of x, y, z coordinates respectively
      */
-    std::array<float, 3> getCoordinates() const;
+    std::array<double, 3> getCoordinates() const;
 
     /**
      * Get material dot velocity as an array
-     * @return Array of floats with copy of vx, vy, vz velocities respectively
+     * @return Array of doubles with copy of vx, vy, vz velocities respectively
      */
-    std::array<float, 3> getVelocity() const;
+    std::array<double, 3> getVelocity() const;
 
     /**
      * Set material dot velocity from vector
-     * @param v vector of floats containing vx, vy, vz to set
+     * @param v vector of doubles containing vx, vy, vz to set
      */
-    void setVelocity(std::array<float, 3> v);
+    void setVelocity(std::array<double, 3> v);
 
     /**
      * Get material dot mass
      * @return Mass of the dot
      */
-    float getMass() const;
+    double getMass() const;
 
     /**
      * Make the dot move in accordance with given <b>force</b> and <b>time</b>
      * @param force Total force that acts on the dot
      * @param t Time period of the force action
      */
-    void evolute(const std::array<float, 3>& force, const float t);
+    void evolute(const std::array<double, 3>& force, const double t);
 
     bool operator==(const MaterialDot& other) const;
